#26 ACTUAL cause of header in column
New errors: header object does not support asynccontextmanager

I am reverting that all back to not try to call async in that same function.

Even those it doesn't always trigger, I needed to pull frame(): out of the main.py the appropriate calls in welcome.py and home_page.py (that's the theory).